:art: Move Movie Details folder to Components folder
