+++ src/Components/MovieDetails/styles.css
